fix: Resolve CI lint failures across multiple crates
- Remove unused import and doc comments in agentfs-interpose-shim
- Fix dangling pointer from temporary in send_fd_open_response
- Add Default implementation for SessionManager in llm-api-proxy
- Fix needless borrowing in proxy.rs and scenario.rs
- Fix unnecessary cast in scenario.rs
- Remove unnecessary cast in tui-testing/src/bin.rs
- Fix derivable Default implementation in ah-rest-client
- Fix unnecessary to_string conversion
- Remove unused variables in debugging-enforcement test
- Fix needless late initialization in ah-macos-launcher
- Fix needless borrows in container-enforcement test

+++ crates/tui-testing/src/bin.rs
@@ -52,7 +52,7 @@ async fn main() -> anyhow::Result<()> {
         .map_err(|e| anyhow::anyhow!("Failed to send request: {}", e))?;
 
     // Receive response with timeout
-    let timeout_duration = Duration::from_secs(args.timeout as u64);
+    let timeout_duration = Duration::from_secs(args.timeout);
     match tokio::time::timeout(timeout_duration, receiver.recv()).await {
         Ok(Ok((response_msg, _))) => {
             let response_bytes = response_msg.iter().next().map(|m| m.as_ref()).unwrap_or(&[][..]);
